Minor: move useAlert.jsx to a new src/hooks folder
It is not a component.
